+++ Makefile
@@ -21,7 +21,7 @@ SRCS		=	$(SRCS_DIR)main.c		\
 				$(SRCS_DIR)tokenize_create_token.c	\
 				$(SRCS_DIR)tokenize_print_token.c	\
 				$(SRCS_DIR)tokenize_utils.c	\
-				$(SRCS_DIR)parser.c	\
+				$(SRCS_DIR)parse.c	\
 				$(SRCS_DIR)vector.c	\
 				$(SRCS_DIR)panic.c	\
 
@@ -31,7 +31,7 @@ SRCS_B		=	$(SRCS_B_DIR)main.c		\
 				$(SRCS_DIR)tokenize_create_token.c	\
 				$(SRCS_DIR)tokenize_print_token.c	\
 				$(SRCS_DIR)tokenize_utils.c	\
-				$(SRCS_DIR)parser.c	\
+				$(SRCS_DIR)parse.c	\
 				$(SRCS_DIR)vector.c	\
 				$(SRCS_DIR)panic.c	\
